Add shared test helpers
