+++ tests/InvestProvider.Backend.Tests/Handlers/MyAllocationHandlerTests.cs
@@ -1,6 +1,4 @@
 using System;
-using System.Collections;
-using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Amazon.DynamoDBv2.DataModel;
@@ -13,56 +11,18 @@ using InvestProvider.Backend.Services.Handlers.MyAllocation;
 using InvestProvider.Backend.Services.Handlers.MyAllocation.Models;
 using Net.Web3.EthereumWallet;
 using FluentValidation;
+using InvestProvider.Backend.Tests;
 
 namespace InvestProvider.Backend.Tests.Handlers;
 
 public class MyAllocationHandlerTests
 {
-    private static object CreatePhase(string id, DateTime start, DateTime finish, decimal maxInvest)
-    {
-        var type = Type.GetType("Poolz.Finance.CSharp.Strapi.ComponentPhaseStartEndAmount, Poolz.Finance.CSharp.Strapi")!;
-        var obj = Activator.CreateInstance(type)!;
-        type.GetProperty("Id")?.SetValue(obj, id);
-        type.GetProperty("Start")?.SetValue(obj, (DateTime?)start);
-        type.GetProperty("Finish")?.SetValue(obj, (DateTime?)finish);
-        var maxInvestProp = type.GetProperty("MaxInvest");
-        if (maxInvestProp != null)
-        {
-            var targetType = Nullable.GetUnderlyingType(maxInvestProp.PropertyType) ?? maxInvestProp.PropertyType;
-            object? converted = Convert.ChangeType(maxInvest, targetType);
-            maxInvestProp.SetValue(obj, converted);
-        }
-        return obj;
-    }
-
-    private static ProjectInfo CreateProjectInfo(long chainId, object phase)
-    {
-        var projectsInfoType = Type.GetType("Poolz.Finance.CSharp.Strapi.ProjectsInformation, Poolz.Finance.CSharp.Strapi")!;
-        var chainSettingType = Type.GetType("Poolz.Finance.CSharp.Strapi.ChainSetting, Poolz.Finance.CSharp.Strapi")!;
-        var chainType = Type.GetType("Poolz.Finance.CSharp.Strapi.Chain, Poolz.Finance.CSharp.Strapi")!;
-
-        var phasesList = (IList)Activator.CreateInstance(typeof(List<>).MakeGenericType(phase.GetType()))!;
-        phasesList.Add(phase);
-
-        var chain = Activator.CreateInstance(chainType)!;
-        chainType.GetProperty("ChainId")?.SetValue(chain, (long?)chainId);
-
-        var chainSetting = Activator.CreateInstance(chainSettingType)!;
-        chainSettingType.GetProperty("Chain")?.SetValue(chainSetting, chain);
-
-        var projectsInfo = Activator.CreateInstance(projectsInfoType)!;
-        projectsInfoType.GetProperty("ChainSetting")?.SetValue(projectsInfo, chainSetting);
-        projectsInfoType.GetProperty("ProjectPhases")?.SetValue(projectsInfo, phasesList);
-
-        var projectInfoResponse = new ProjectInfoResponse((dynamic)projectsInfo);
-        return new ProjectInfo(projectInfoResponse);
-    }
 
     [Fact]
     public async Task Handle_ReturnsResponse_WhenDataExists()
     {
-        var phase = CreatePhase("1", DateTime.UtcNow, DateTime.UtcNow.AddHours(1), 0m);
-        var projectInfo = CreateProjectInfo(1, phase);
+        var phase = TestHelpers.CreatePhase("1", DateTime.UtcNow, DateTime.UtcNow.AddHours(1), 0m);
+        var projectInfo = TestHelpers.CreateProjectInfo(1, phase);
 
         var strapi = new Mock<IStrapiClient>();
         strapi.Setup(x => x.ReceiveProjectInfo("pid", false)).Returns(projectInfo);
@@ -94,8 +54,8 @@ public class MyAllocationHandlerTests
     [Fact]
     public async Task Handle_Throws_WhenWhiteListMissing()
     {
-        var phase = CreatePhase("2", DateTime.UtcNow, DateTime.UtcNow.AddHours(1), 0m);
-        var projectInfo = CreateProjectInfo(1, phase);
+        var phase = TestHelpers.CreatePhase("2", DateTime.UtcNow, DateTime.UtcNow.AddHours(1), 0m);
+        var projectInfo = TestHelpers.CreateProjectInfo(1, phase);
 
         var strapi = new Mock<IStrapiClient>();
         strapi.Setup(x => x.ReceiveProjectInfo("pid", false)).Returns(projectInfo);
